Refactor: SingleModePage 로직 분리
