+++ src/pages/singlemode/hooks/usePageProtection.ts
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+
+export const usePageProtection = () => {
+  const preventClose = (e: BeforeUnloadEvent) => {
+    e.preventDefault();
+  };
+
+  useEffect(() => {
+    window.addEventListener("beforeunload", preventClose);
+    
+    return () => {
+      window.removeEventListener("beforeunload", preventClose);
+    };
+  }, []);
+};
