+++ src/pages/singlemode/hooks/useTopicInitialization.ts
@@ -1,0 +1,10 @@
+import { useEffect } from "react";
+import { useDrawingStore } from "../../../stores/drawingStore";
+
+export const useTopicInitialization = () => {
+  const { getRandomTopic } = useDrawingStore();
+
+  useEffect(() => {
+    getRandomTopic();
+  }, [getRandomTopic]);
+};
